Addressed several issues
